Готовая др...

+++ src/main/java/web/dao/UserDao.java
@@ -7,6 +7,6 @@ import java.util.List;
 public interface UserDao {
     List<User> getAll();
     void save(User user);
-    User getById(int id);
-    void delete(int id);
+    User getById(Long id);
+    void delete(Long id);
 }
