+++ src/main/java/web/service/UserServiceImpl.java
@@ -27,13 +27,13 @@ public class UserServiceImpl implements UserService {
 
     @Override
     @Transactional(readOnly = true)
-    public User getById(int id) {
+    public User getById(Long id) {
         return userDao.getById(id);
     }
 
     @Override
     @Transactional
-    public void delete(int id) {
+    public void delete(Long id) {
         userDao.delete(id);
     }
 }
